Trunk-BP: $Id$-Tag, include <sys/types.h>, fixed a prototype.

+++ include/atalk/netddp.h
@@ -1,4 +1,6 @@
-/* 
+/*
+ * $Id: netddp.h,v 1.1.1.1.2.1 2002-02-08 00:01:08 srittau Exp $
+ *
  * Copyright (c) 1999 Adrian Sun (asun@zoology.washington.edu)
  * All Rights Reserved. See COPYRIGHT.
  *
@@ -11,6 +13,7 @@
 #ifndef _ATALK_NETDDP_H
 #define _ATALK_NETDDP_H 1
 
+#include <sys/types.h>
 #include <sys/cdefs.h>
 #include <sys/socket.h>
 #include <netatalk/at.h>
@@ -18,7 +21,7 @@
 extern int netddp_open   __P((struct sockaddr_at *, struct sockaddr_at *));
 
 #if !defined(NO_DDP) && defined(MACOSX_SERVER)
-extern int netddp_sendto __P((int, void *, int, unsigned int, 
+extern int netddp_sendto __P((int, void *, size_t, unsigned int, 
 			   const struct sockaddr *, unsigned int));
 extern int netddp_recvfrom __P((int, void *, int, unsigned int, 
 			     struct sockaddr *, unsigned int *));
